challs_finished/Yugen's Choice/secure_coding_yugens/challenge/app/requirements.txt
Flask
redis
python-dotenv
Werkzeug
